Update Vitrovex scintillation properties to match DVT measurements in air
Commits e088a5d and c87a047 matched simulation rates to ice instead of air, which resulted in an overestimation of absolute rates.

This bypasses the method of measured yields and mirrors measurements exactly.

This will most likely be reverted once the initial yield matching is redone for the correct environment.